modified functions to pass new test cases.
Changed driver to write-only, before it was read/write with memory address, which would allow the test to change the memory address and the driver would read that change and write to that modified value instead of writing to the last known memory address. 

added all on function, added boundary condition to led_TurnOn and led_TurnOff to avoid odd behavior caused by out of bounds errors.